% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/angle_laplace.R
\name{angle_laplace}
\alias{angle_laplace}
\title{Draws a random turning angle from a truncated Laplace distribution.}
\usage{
angle_laplace(scale = 1)
}
\arguments{
\item{scale}{scale parameter for the truncated Laplace distribution}
}
\value{
random angle (radians).
}
\description{
Draws a random turning angle in the range (-pi, pi] from a truncated
Laplace distribution with the given scale. This is a short-cut for
\code{rtrunclaplace(1, 0, scale, -pi, pi)}.
}
\seealso{
\code{\link{rtrunclaplace}}
}

