% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{Laplace}
\alias{Laplace}
\alias{dlaplace}
\alias{plaplace}
\alias{rlaplace}
\title{The Laplace distribution}
\usage{
dlaplace(x, mean, scale)

plaplace(q, mean, scale)

rlaplace(n, mean, scale)
}
\arguments{
\item{x,}{q vector of quantiles}

\item{mean}{vector of means}

\item{scale}{vector of scale parameters (> 0)}

\item{n}{number of observations. If length(n) > 1, the length is taken
to be the number required}
}
\description{
Density, distribution function, quantile function and random generation for the
Laplace distribution.
}

