% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_walk.R
\name{plot_walk}
\alias{plot_walk}
\title{Plots walk paths using ggplot.}
\usage{
plot_walk(w, colour = c("none", "direct", "path"), legend = FALSE,
  by.id = FALSE, nrow = NULL, ncol = NULL)
}
\arguments{
\item{w}{Either a single \code{walk.data} object, or a list of
such objects.}

\item{colour}{The variable to use for colouring within each walk path:
\code{'direct'} for direct distance to start point;
\code{'path'} for distance along path;
\code{'none'} for no colouring}

\item{legend}{If \code{TRUE}, a legend is added to the plot when path
colouring is used. Ignored if \code{colour} is \code{'none'}.}

\item{by.id}{If \code{TRUE}, plots are facetted by the \code{walk.data$id}
field. If \code{FALSE} (default), a single plot is produced.}

\item{nrow}{If \code{by.id} is \code{TRUE}, this argument can be used to set the
number of rows of facetted plots.}

\item{ncol}{If \code{by.id} is \code{TRUE}, this argument can be used to set the
  number of columns of facetted plots. See \code{link{ggplot2::facet_wrap}}.}
}
\value{
A plot object.
}
\description{
Plots one or more \code{walk.data} object paths (generated by
\code{\link{make_walk}}).
}

