% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trunc_laplace_distribution.R
\name{Laplace (truncated)}
\alias{Laplace (truncated)}
\alias{dtrunclaplace}
\alias{ptrunclaplace}
\alias{rtrunclaplace}
\title{The truncated Laplace distribution}
\usage{
dtrunclaplace(x, mean, scale, lower, upper)

ptrunclaplace(q, mean, scale, lower, upper)

rtrunclaplace(n, mean, scale, lower, upper)
}
\arguments{
\item{x,}{q vector of quantiles}

\item{mean}{vector of means}

\item{scale}{vector of scale parameters (> 0)}

\item{lower}{lower bound of distribution}

\item{upper}{upper bound of distribution}

\item{n}{number of observations. If length(n) > 1, the length is taken
to be the number required}
}
\description{
Density, distribution function and random generation for the truncated
Laplace distribution.
}
\examples{
\dontrun{
# Compare the distribution of random angles from a truncated
# Laplace versus those from a standard Laplace distribution.
N <- 1000
as <- rlaplace(N, mean=0 scale=1.0)
at <- rtrunclaplace(N, mean=0, scale=1.0, lower=-pi/2, upper=pi/2)

library(ggplot2)
dat <- data.frame(angle = c(as, at), distribution = rep(c("standard", "truncated"), each=N))
ggplot() +
  geom_histogram(data=dat, aes(x=angle, fill=distribution), position="dodge", binwidth=pi/10) +
  theme_bw()
}
}

